Move store and reducers to action folder
