+++ web/reactjs/src/actions/store.js
@@ -12,27 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React from 'react';
-import Box from '@material-ui/core/Box';
+import 'regenerator-runtime/runtime';
+import { applyMiddleware, createStore, combineReducers } from 'redux';
+import { createLogger } from 'redux-logger';
+import { createPromise } from 'redux-promise-middleware';
+import { authReducer } from './reducers';
 
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles(theme => ({
-    root: {
-        padding: theme.spacing(4)
-    }
-}));
-
-const DashboardItem = () => {
-
-    const classes = useStyles();
-
-    return (
-        <div className={classes.root}>
-            <Box container spacing={1}>
-            </Box>
-        </div>
-    );
-};
-  
-export default DashboardItem;
+export const store = createStore(
+    combineReducers({authReducer}),
+    {},
+    applyMiddleware(createPromise({types:{fulfilled: 'success'}}), createLogger())
+);
