+++ web/reactjs/src/modules/Dashboard/components/index.js
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export { default as authReducer } from './authReducer';
+export { default } from './Dashboard';
